+++ changedetectionio/blueprint/browser_notifications/__init__.py
@@ -1,0 +1,1 @@
+# Browser notifications blueprint
